refactor: organize DTO interfaces
