+++ src/application/contracts/usecase.interface.ts
@@ -1,21 +1,7 @@
-export interface InitBaileysInstanceUseCaseInputDTO {
-    key: string;
-    webhook?: string;
-    allowWebhook?: boolean;
-    heardEvents: string[];
-    isWebhookBase64?: boolean;
-    markMessagesRead?: boolean;
-    apiKey?: string;
-}
-
-export interface InitBailesInstanceUseCaseOutputDTO {
-    key: string;
-    webhook: string | undefined;
-    allowWebhook: boolean;
-    heardEvents: string[];
-    isOn: boolean;
-
-}
+import {
+    InitBailesInstanceUseCaseOutputDTO,
+    InitBaileysInstanceUseCaseInputDTO
+} from "../features/commands/baileys/init-baileys-instance/init-baileys-instance.dto";
 
 export interface InitBailesInstanceUseCaseInterface {
     execute(input: InitBaileysInstanceUseCaseInputDTO): Promise<InitBailesInstanceUseCaseOutputDTO>;
