+++ src/application/features/commands/baileys/init-baileys-instance/init-baileys-instance.dto.ts
@@ -1,0 +1,18 @@
+export interface InitBaileysInstanceUseCaseInputDTO {
+    key: string;
+    webhook?: string;
+    allowWebhook?: boolean;
+    heardEvents: string[];
+    isWebhookBase64?: boolean;
+    markMessagesRead?: boolean;
+    apiKey?: string;
+}
+
+export interface InitBailesInstanceUseCaseOutputDTO {
+    key: string;
+    webhook: string | undefined;
+    allowWebhook: boolean;
+    heardEvents: string[];
+    isOn: boolean;
+
+}
